Move Avatar to design system
