+++ web/src/components/design-system/Avatar/Avatar.tsx
