+++ web/src/features/comments/CommentList.tsx
@@ -1,5 +1,5 @@
 /* eslint-disable @repo/no-style-props */
-import { Avatar } from "@/src/components/ui/avatar";
+import { Avatar } from "@/src/components/design-system/Avatar/Avatar";
 import { Button } from "@/src/components/ui/button";
 import { KeyboardShortcut } from "@/src/components/design-system/KeyboardShortcut/KeyboardShortcut";
 import {
